+++ src/resource.ts
@@ -1,15 +1,13 @@
 import { mount, unmount } from './allocation';
 import { ownership } from './state';
+import { MountableResource } from './types';
 
 /**
  * Represents an arbitrary stateful resource that is asynchronously provisioned
  * and destroyed. Resources can own other resources, and destroying a parent
  * first tears down the children.
  */
-export default abstract class Resource<
-  Controls extends object,
-  Args extends Array<unknown> = [],
-> {
+export default abstract class Resource<Controls extends object> {
   #resources = new WeakSet<object>();
   #children: Array<object> = [];
 
@@ -17,26 +15,17 @@ export default abstract class Resource<
     ownership.set(this, this.#children);
   }
 
-  /** A hook that gets called when the resource is created. */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async enter(..._args: Args) {
-    return;
-  }
-
-  /** A hook that gets called when the resource is destroyed. */
-  async leave() {
-    return;
-  }
-
   /** Provision an owned resource and make sure it doesn't outlive us. */
   protected async allocate<
     ChildControls extends object,
-    ChildArgs extends Array<unknown> = [],
+    ChildArgs extends Array<unknown>,
   >(
-    Child: new () => Resource<ChildControls, ChildArgs>,
+    Child: new () =>
+      | MountableResource<ChildControls, ChildArgs>
+      | Resource<ChildControls>,
     ...args: ChildArgs
   ): Promise<ChildControls> {
-    const controls = await mount(Child, ...args);
+    const controls: ChildControls = await mount(Child, ...args);
     this.#resources.add(controls);
     this.#children.push(controls);