+++ src/types.ts
@@ -1,11 +1,8 @@
 import type Resource from './resource';
 
 /** The `exports` type for a resource. */
-export type Controls<
-  ArbitraryResource extends Resource<object, Array<unknown>>,
-> = ArbitraryResource extends Resource<infer Controls, Array<unknown>>
-  ? Controls
-  : never;
+export type Controls<ArbitraryResource extends Resource<object>> =
+  ArbitraryResource extends Resource<infer Controls> ? Controls : never;
 
 export interface MountableResource<
   Controls extends object,