+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-06-04
+
 ### Changed
 
 - Prevent use of `allocate(...)`/`deallocate(...)` outside a resource subclass.
@@ -40,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `mount`/`unmount` hooks to provision resources
 - `allocate`/`deallocate` for creating hierarchies of resources
 
-[Unreleased]: https://github.com/PsychoLlama/wardens/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/PsychoLlama/wardens/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/PsychoLlama/wardens/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/PsychoLlama/wardens/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/PsychoLlama/wardens/releases/tag/v0.1.0